+++ selectrum.el
@@ -480,9 +480,6 @@ Passed to various hook functions.")
 (defvar selectrum--count-overlay nil
   "Overlay used to display count information before prompt.")
 
-(defvar selectrum--right-margin-overlays nil
-  "A list of overlays used to display right margin text.")
-
 (defvar selectrum--last-command nil
   "Name of last interactive command that invoked Selectrum.")
 
@@ -657,7 +654,11 @@ PRED defaults to `minibuffer-completion-predicate'."
                                     (setq input (or (alist-get 'input result)
                                                     input))
                                     (setq selectrum--visual-input input)
-                                    (alist-get 'candidates result))))
+                                    ;; Avoid modifying the returned
+                                    ;; candidates to let the function
+                                    ;; reuse them.
+                                    (copy-sequence
+                                     (alist-get 'candidates result)))) )
                      selectrum--preprocessed-candidates)))
         (setq selectrum--total-num-candidates (length cands))
         (setq selectrum--refined-candidates
@@ -682,7 +683,8 @@ PRED defaults to `minibuffer-completion-predicate'."
         (setq selectrum--current-candidate-index
               (cond
                ((null selectrum--refined-candidates)
-                nil)
+                (when (not selectrum--match-required-p)
+                  -1))
                ((and selectrum--default-candidate
                      (string-empty-p (minibuffer-contents))
                      (not (member selectrum--default-candidate
@@ -709,8 +711,6 @@ PRED defaults to `minibuffer-completion-predicate'."
                  'before-string (selectrum--count-info))
     (overlay-put selectrum--count-overlay
                  'priority 1)
-    (while selectrum--right-margin-overlays
-      (delete-overlay (pop selectrum--right-margin-overlays)))
     (setq input (or selectrum--visual-input input))
     (let* ((first-index-displayed
             (if selectrum--current-candidate-index
@@ -737,11 +737,6 @@ PRED defaults to `minibuffer-completion-predicate'."
       (setq displayed-candidates
             (seq-take displayed-candidates
                       selectrum-num-candidates-displayed))
-      (let ((n (1+ (if selectrum-fix-minibuffer-height
-                       selectrum-num-candidates-displayed
-                     (max (1- (window-height)) ; grow only
-                          (length displayed-candidates))))))
-        (setf (window-height) n))
       (let ((text (selectrum--candidates-display-string
                    displayed-candidates
                    input
@@ -774,9 +769,8 @@ PRED defaults to `minibuffer-completion-predicate'."
                                    'selectrum-current-candidate
                                  'minibuffer-prompt))
                               (propertize "]" 'face 'minibuffer-prompt)))
-              (unless (or (and highlighted-index
-                               (>= highlighted-index 0))
-                          selectrum--match-required-p)
+              (when (and highlighted-index
+                         (< highlighted-index 0))
                 (add-text-properties
                  (minibuffer-prompt-end) bound
                  '(face selectrum-current-candidate))))
@@ -787,13 +781,44 @@ PRED defaults to `minibuffer-completion-predicate'."
                       (point-max) (point-max) (current-buffer))
         (setq text (concat (or default " ") text))
         (put-text-property 0 1 'cursor t text)
-        (overlay-put selectrum--candidates-overlay 'after-string text)))
+        (overlay-put selectrum--candidates-overlay 'after-string text))
+      (selectrum--update-minibuffer-height first-index-displayed
+                                           highlighted-index
+                                           displayed-candidates))
     (setq selectrum--end-of-input-marker (set-marker (make-marker) bound))
     (set-marker-insertion-type selectrum--end-of-input-marker t)
-    (selectrum--fix-set-minibuffer-message)
     (when keep-mark-active
       (setq deactivate-mark nil))
     (setq-local selectrum--init-p nil)))
+
+(defun selectrum--update-minibuffer-height (first highlighted cands)
+  "Set minibuffer height for candidates display.
+FIRST is the index of the first displayed candidate. HIGHLIGHTED
+is the index if the highlighted candidate. CANDS are the
+currently displayed candidates."
+  (when-let ((n (1+ selectrum-num-candidates-displayed))
+             (win (active-minibuffer-window)))
+    ;; Don't attempt to resize a minibuffer frame.
+    (unless (frame-root-window-p win)
+      ;; Set min initial height.
+      (when (and selectrum-fix-minibuffer-height
+                 selectrum--init-p)
+        (with-selected-window win
+          (setf (window-height) n)))
+      ;; Adjust if needed.
+      (when (or selectrum--init-p
+                (and selectrum--current-candidate-index
+                     ;; Allow size change when navigating, not while
+                     ;; typing.
+                     (/= first highlighted)
+                     ;; Don't allow shrinking.
+                     (= (length cands)
+                        selectrum-num-candidates-displayed)))
+        (let ((dheight (cdr (window-text-pixel-size win)))
+              (wheight (window-pixel-height win)))
+          (when (/= dheight wheight)
+            (window-resize
+             win (- dheight wheight) nil nil 'pixelwise)))))))
 
 (defun selectrum--first-lines (candidates)
   "Return list of single line CANDIDATES.
@@ -925,18 +950,23 @@ candidate."
                 'minibuffer-prompt))))
           (insert displayed-candidate)
           (when right-margin
-            (let ((ol (make-overlay (point) (point))))
-              (overlay-put
-               ol 'after-string
-               (concat
-                (propertize
-                 " "
-                 'display
-                 `(space :align-to (- right-fringe
-                                      ,(string-width right-margin)
-                                      selectrum-right-margin-padding)))
-                right-margin))
-              (push ol selectrum--right-margin-overlays))))
+            (insert
+             (concat
+              (propertize
+               " "
+               'face
+               (when (and right-margin
+                          (equal index highlighted-index))
+                 'selectrum-current-candidate)
+               'display
+               `(space :align-to (- right-fringe
+                                    ,(string-width right-margin)
+                                    selectrum-right-margin-padding)))
+              (propertize right-margin
+                          'face
+                          (when (and right-margin
+                                     (equal index highlighted-index))
+                            'selectrum-current-candidate))))))
         (cl-incf index))
       (buffer-string))))
 
@@ -947,9 +977,7 @@ candidate."
   (remove-hook 'minibuffer-exit-hook #'selectrum--minibuffer-exit-hook 'local)
   (when (overlayp selectrum--count-overlay)
     (delete-overlay selectrum--count-overlay))
-  (setq selectrum--count-overlay nil)
-  (while selectrum--right-margin-overlays
-    (delete-overlay (pop selectrum--right-margin-overlays))))
+  (setq selectrum--count-overlay nil))
 
 (cl-defun selectrum--minibuffer-setup-hook
     (candidates &key default-candidate initial-input)
@@ -961,8 +989,9 @@ into the user input area to start with."
   (add-hook
    'minibuffer-exit-hook #'selectrum--minibuffer-exit-hook nil 'local)
   (setq-local selectrum--init-p t)
-  (setq selectrum--candidates-overlay
-        (make-overlay (point) (point) nil 'front-advance 'rear-advance))
+  (unless selectrum--candidates-overlay
+    (setq selectrum--candidates-overlay
+          (make-overlay (point) (point) nil 'front-advance 'rear-advance)))
   (setq selectrum--start-of-input-marker (point-marker))
   (if selectrum--repeat
       (insert selectrum--previous-input-string)
@@ -1142,7 +1171,8 @@ list). A null or non-positive ARG inserts the candidate corresponding to
             (or (get-text-property 0 'selectrum--candidate-insert candidate)
                 (selectrum--get-full candidate))))
       (insert full)
-      (add-to-history minibuffer-history-variable full)
+      (unless (eq t minibuffer-history-variable)
+        (add-to-history minibuffer-history-variable full))
       (apply
        #'run-hook-with-args
        'selectrum-candidate-inserted-hook
@@ -1204,7 +1234,6 @@ Otherwise, just eval BODY."
               selectrum--visual-input
               selectrum--read-args
               selectrum--count-overlay
-              selectrum--right-margin-overlays
               selectrum--repeat
               selectrum-active-p)))
      ;; https://github.com/raxod502/selectrum/issues/39#issuecomment-618350477
@@ -1274,7 +1303,8 @@ and `minibuffer-completion-predicate'. They are used for internal
 purposes and compatibility to Emacs completion API. By passing
 these as keyword arguments they will be dynamically bound as per
 semantics of `cl-defun'."
-  (unless may-modify-candidates
+  (unless (or may-modify-candidates
+              (functionp candidates))
     (setq candidates (copy-sequence candidates)))
   (selectrum--save-global-state
     (setq selectrum--read-args (cl-list* prompt candidates args))
@@ -1304,20 +1334,6 @@ semantics of `cl-defun'."
                (max-mini-window-height
                 (1+ selectrum-num-candidates-displayed))
                (prompt (selectrum--remove-default-from-prompt prompt))
-               ;; Need to bind this back to its standard value due to
-               ;; <https://github.com/raxod502/selectrum/issues/61>.
-               ;; What happens is `selectrum-read-file-name' binds
-               ;; `completing-read-function' to
-               ;; `selectrum--completing-read-file-name', so if you
-               ;; invoke another Selectrum command recursively then it
-               ;; inherits that binding, even if the new Selectrum
-               ;; command is not reading file names. This causes an
-               ;; error. Arguably this solution is a bit of a hack but
-               ;; it should work "well enough" for now. If we
-               ;; encounter more trouble then we shall come up with a
-               ;; proper solution.
-               (completing-read-function
-                #'selectrum-completing-read)
                ;; <https://github.com/raxod502/selectrum/issues/99>
                (icomplete-mode nil)
                (selectrum-active-p t))
@@ -1610,28 +1626,31 @@ less appropriate. It also allows you to view hidden buffers,
 which is otherwise impossible due to tricky behavior of Emacs'
 completion machinery. For PROMPT, DEF, REQUIRE-MATCH, and
 PREDICATE, see `read-buffer'."
-  (let ((selectrum-should-sort-p nil)
-        (candidates
-         (lambda (input)
-           (let* ((buffers (mapcar #'buffer-name (buffer-list)))
-                  (candidates (if predicate
-                                  (cl-delete-if-not predicate buffers)
-                                buffers)))
-             (if (string-prefix-p " " input)
-                 (progn
-                   (setq input (substring input 1))
-                   (setq candidates
-                         (cl-delete-if-not
-                          (lambda (name)
-                            (string-prefix-p " " name))
-                          candidates)))
-               (setq candidates
-                     (cl-delete-if
-                      (lambda (name)
-                        (string-prefix-p " " name))
-                      candidates)))
-             `((candidates . ,candidates)
-               (input . ,input))))))
+  (let* ((selectrum-should-sort-p nil)
+         (buffalist (mapcar (lambda (buf)
+                              (cons (buffer-name buf) buf))
+                            (buffer-list)))
+         (buffers (mapcar #'car (if predicate
+                                    (cl-delete-if-not predicate buffalist)
+                                  buffalist)))
+         (candidates
+          (lambda (input)
+            (let ((candidates (copy-sequence buffers)))
+              (if (string-prefix-p " " input)
+                  (progn
+                    (setq input (substring input 1))
+                    (setq candidates
+                          (cl-delete-if-not
+                           (lambda (name)
+                             (string-prefix-p " " name))
+                           candidates)))
+                (setq candidates
+                      (cl-delete-if
+                       (lambda (name)
+                         (string-prefix-p " " name))
+                       candidates)))
+              `((candidates . ,candidates)
+                (input . ,input))))))
     (substring-no-properties
      (selectrum-read
       prompt candidates
@@ -1707,13 +1726,57 @@ For PROMPT, COLLECTION, PREDICATE, REQUIRE-MATCH, INITIAL-INPUT,
   "Read file name using Selectrum. Can be used as `read-file-name-function'.
 For PROMPT, DIR, DEFAULT-FILENAME, MUSTMATCH, INITIAL, and
 PREDICATE, see `read-file-name'."
-  (let ((completing-read-function #'selectrum--completing-read-file-name))
-    (minibuffer-with-setup-hook
-        (:append (lambda ()
-                   (set-syntax-table
-                    selectrum--minibuffer-local-filename-syntax)))
+  (minibuffer-with-setup-hook
+      (:append (lambda ()
+                 (when (and default-filename
+                            ;; ./ should be omitted.
+                            (not (equal
+                                  (expand-file-name default-filename)
+                                  (expand-file-name default-directory))))
+                   (setq selectrum--default-candidate
+                         ;; Sort for directories needs any final
+                         ;; slash removed.
+                         (directory-file-name
+                          ;; The candidate should be sorted by it's
+                          ;; relative name.
+                          (file-relative-name default-filename
+                                              default-directory))))
+                 (set-syntax-table
+                  selectrum--minibuffer-local-filename-syntax)))
+    (let* ((crf completing-read-function)
+           ;; See <https://github.com/raxod502/selectrum/issues/61>.
+           ;; When you invoke another `completing-read' command
+           ;; recursively then it inherits the
+           ;; `completing-read-function' binding, and unless it's
+           ;; another file reading command using
+           ;; `selectrum--completing-read-file-name' this will cause
+           ;; an error. To circumvent this we use the function to
+           ;; reset the variable when called.
+           (completing-read-function
+            (lambda (&rest args)
+              (setq completing-read-function crf)
+              (apply #'selectrum--completing-read-file-name args))))
       (read-file-name-default
-       prompt dir default-filename mustmatch initial predicate))))
+       prompt dir
+       ;; We don't pass default-candidate here to avoid that
+       ;; submitting the selected prompt results in the default file
+       ;; name. This is the stock Emacs behavior where there is no
+       ;; concept of an active selection. Instead we pass the initial
+       ;; prompt as default so it gets returned when submitted. In
+       ;; addition to that we set `selectrum--default-candidate' in
+       ;; the setup hook above so the actual default gets sorted to
+       ;; the top. This should give the same convenience as in default
+       ;; completion (where you can press RET at the initial prompt to
+       ;; get the default). The downside is that this convenience is
+       ;; gone when sorting is disabled or the default-filename is
+       ;; outside the prompting directory but this should be rare
+       ;; case.
+       (concat
+        (expand-file-name
+         (or dir
+             default-directory))
+        initial)
+       mustmatch initial predicate))))
 
 (defvar selectrum--old-read-file-name-function nil
   "Previous value of `read-file-name-function'.")
@@ -1834,53 +1897,25 @@ shadows correctly."
     (call-interactively selectrum--last-command)))
 
 ;;;###autoload
-(defun selectrum--fix-set-minibuffer-message (&rest _)
-  "Move the minibuffer message overlay to the right place.
-This advice fixes the overlay placed by `set-minibuffer-message',
-which is different from the one placed by `minibuffer-message'.
-
-By default the overlay is placed at the end, but in the case of
-Selectrum this means after all the candidates. We want to move it
-instead to just after the user input.
-
-To test that this advice is working correctly, type \\[find-file]
-and enter \"/sudo::\", then authenticate. The overlay indicating
-that authentication was successful should appear right after the
-user input area, not at the end of the candidate list.
-
-This is an `:after' advice for `set-minibuffer-message'."
-  (selectrum--when-compile (boundp 'minibuffer-message-overlay)
-    (when (and (bound-and-true-p selectrum-active-p)
-               (overlayp minibuffer-message-overlay))
-      (move-overlay minibuffer-message-overlay
-                    selectrum--end-of-input-marker
-                    selectrum--end-of-input-marker))))
-
-;;;###autoload
 (defun selectrum--fix-minibuffer-message (func &rest args)
-  "Move the minibuffer message overlay to the right place.
-This advice fixes the overlay placed by `minibuffer-message',
-which is different from the one placed by
-`set-minibuffer-message'.
-
-By default the overlay is placed at the end, but in the case of
-Selectrum this means after all the candidates. We want to move it
-instead to just after the user input.
+  "Ensure the cursor stays at the front of the minibuffer message.
+This advice adjusts where the cursor gets placed for the overlay
+of `minibuffer-message'.
 
 To test that this advice is working correctly, type \\[find-file]
 twice in a row. The overlay indicating that recursive minibuffers
 are not allowed should appear right after the user input area,
-not at the end of the candidate list.
+not at the end of the candidate list and the cursor should stay
+at the front.
 
 This is an `:around' advice for `minibuffer-message'. FUNC and
 ARGS are standard as in all `:around' advice."
   (if (bound-and-true-p selectrum-active-p)
-      (cl-letf* ((orig-make-overlay (symbol-function #'make-overlay))
-                 ((symbol-function #'make-overlay)
-                  (lambda (_beg _end &rest args)
-                    (apply orig-make-overlay
-                           selectrum--end-of-input-marker
-                           selectrum--end-of-input-marker
+      (cl-letf* ((orig-put-text-property (symbol-function #'put-text-property))
+                 ((symbol-function #'put-text-property)
+                  (lambda (beg end key val &rest args)
+                    (apply orig-put-text-property
+                           beg end key (if (eq key 'cursor) 1 val)
                            args))))
         (apply func args))
     (apply func args)))
@@ -1929,10 +1964,6 @@ ARGS are standard as in all `:around' advice."
                       #'selectrum-read-library-name)
           (advice-add #'minibuffer-message :around
                       #'selectrum--fix-minibuffer-message)
-          ;; No sharp quote because `set-minibuffer-message' is not
-          ;; defined in older Emacs versions.
-          (advice-add 'set-minibuffer-message :after
-                      #'selectrum--fix-set-minibuffer-message)
           (define-key minibuffer-local-map
             [remap previous-matching-history-element]
             'selectrum-select-from-history))
@@ -1961,10 +1992,6 @@ ARGS are standard as in all `:around' advice."
       ;; older Emacs versions.
       (advice-remove 'read-library-name #'selectrum-read-library-name)
       (advice-remove #'minibuffer-message #'selectrum--fix-minibuffer-message)
-      ;; No sharp quote because `set-minibuffer-message' is not
-      ;; defined in older Emacs versions.
-      (advice-remove 'set-minibuffer-message
-                     #'selectrum--fix-set-minibuffer-message)
       (when (eq (lookup-key minibuffer-local-map
                             [remap previous-matching-history-element])
                 #'selectrum-select-from-history)