+++ CHANGELOG.md
@@ -35,6 +35,12 @@ The format is based on [Keep a Changelog].
   [#113] and [#118].
 
 ### Enhancements
+* If `selectrum-candidate-display-right-margin` is used the margin is
+  included in the highlighting of the selected candidate so it's
+  easier to see to which candidate the margin belongs to.
+* If the `default-filename` passed to `selectrum-read-file-name` is an
+  absolute path it will still be sorted to the top when it is
+  contained in the prompting directory ([#160]).
 * `icomplete-mode` is now automatically disabled when entering
   Selectrum, to avoid conflicts ([#99]).
 * Working with the default candidate has been improved in cases where
@@ -82,6 +88,30 @@ The format is based on [Keep a Changelog].
   ([#136], [#156]).
 * `org-set-tags-command` completion wasn't working after the first
   inserted tag which has been fixed ([#139], [#156]).
+* The mininbuffer height is now determined by the actual height of
+  displayed candidates. Previously the height could be off for
+  candidates containing unicode characters or other means which
+  changed the display height such as `line-spacing` ([#146], [#151],
+  [#154]).
+* When passing a named function or compiled lambda as `candidates`
+  argument to `selectrum-read` an error would be thrown, which has
+  been fixed ([#163]).
+* When selecting file name prompts and submitting them the result will
+  be the selected prompt. Previously it could be the default file name
+  passed to `selectrum-read-file-name` which is the behavior of
+  default completion where there is no concept of an active selection
+  ([#157], [#160]).
+* When there are no candidates and a match isn't required the prompt
+  will now be shown as selected to indicate one can submit the input
+  ([#160]).
+* If the `mustmatch` argument to `read-directory-name` was non-nil the
+  selection of the prompt wasn't visible which has been fixed ([#157],
+  [#160]).
+* If a predicate was passed to `read-buffer` an error would be thrown
+  which has been fixed ([#159], [#161]).
+* Dynamic collection functions can now reuse their returned
+  candidates. Previously Selectrum could modify them even when the
+  `:may-modify-candidates` argument wasn't passed to `selectrum-read`.
 * Incremental history search via `isearch` wasn't working which has
   been fixed ([#124], [#133]).
 * Empty string completion candidates are now ignored like in the
@@ -97,6 +127,10 @@ The format is based on [Keep a Changelog].
   `completing-read-multiple` when `crm-separator` has a non default
   value. Previously it would replace the separator with commas when
   adding new candidates ([#140]).
+* `selectrum-insert-current-candidate` now works from
+  `selectrum-select-from-history` and other commands which ignore
+  history by setting `minibuffer-history-variable` to `t`. Previously
+  an error would be thrown ([#152]).
 
 [#67]: https://github.com/raxod502/selectrum/issues/67
 [#82]: https://github.com/raxod502/selectrum/issues/82
@@ -122,7 +156,19 @@ The format is based on [Keep a Changelog].
 [#138]: https://github.com/raxod502/selectrum/pull/138
 [#139]: https://github.com/raxod502/selectrum/issues/139
 [#140]: https://github.com/raxod502/selectrum/pull/140
+<<<<<<< HEAD
 [#156]: https://github.com/raxod502/selectrum/pull/156
+=======
+[#146]: https://github.com/raxod502/selectrum/issues/146
+[#151]: https://github.com/raxod502/selectrum/issues/151
+[#152]: https://github.com/raxod502/selectrum/pull/152
+[#154]: https://github.com/raxod502/selectrum/pull/154
+[#157]: https://github.com/raxod502/selectrum/issues/157
+[#159]: https://github.com/raxod502/selectrum/issues/159
+[#160]: https://github.com/raxod502/selectrum/pull/160
+[#161]: https://github.com/raxod502/selectrum/pull/161
+[#163]: https://github.com/raxod502/selectrum/pull/163
+>>>>>>> master
 
 ## 2.0 (released 2020-07-18)
 ### Breaking changes